refactor: move All component class functionality to service class and remove All component class
