+++ src/main/java/subway/section/service/SectionService.java
@@ -2,17 +2,16 @@ package subway.section.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.line.component.LineComponent;
 import subway.line.entity.Line;
-import subway.section.component.SectionComponent;
+import subway.line.service.LineService;
 import subway.section.dto.SectionRequest;
 import subway.section.dto.SectionResponse;
 import subway.section.entity.Section;
 import subway.section.entity.Sections;
 import subway.section.exception.SectionException;
 import subway.section.repository.SectionRepository;
-import subway.station.component.StationComponent;
 import subway.station.entity.Station;
+import subway.station.service.StationService;
 
 import java.util.List;
 
@@ -23,24 +22,22 @@ import static subway.converter.LineConverter.convertToStationIds;
 public class SectionService {
 
     private final SectionRepository sectionRepository;
-    private final StationComponent stationComponent;
-    private final LineComponent lineComponent;
-    private final SectionComponent sectionComponent;
+    private final StationService stationService;
+    private final LineService lineService;
 
-    public SectionService(SectionRepository sectionRepository, StationComponent stationComponent, LineComponent lineComponent, SectionComponent sectionComponent) {
+    public SectionService(SectionRepository sectionRepository, StationService stationService, LineService lineService) {
         this.sectionRepository = sectionRepository;
-        this.stationComponent = stationComponent;
-        this.lineComponent = lineComponent;
-        this.sectionComponent = sectionComponent;
+        this.stationService = stationService;
+        this.lineService = lineService;
     }
 
     @Transactional
     public SectionResponse createSection(Long lineId, SectionRequest sectionRequest) {
-        Station upStation = stationComponent.getStationByIdOrThrow(sectionRequest.getUpStationId());
-        Station downStation = stationComponent.getStationByIdOrThrow(sectionRequest.getDownStationId());
+        Station upStation = stationService.getStationByIdOrThrow(sectionRequest.getUpStationId());
+        Station downStation = stationService.getStationByIdOrThrow(sectionRequest.getDownStationId());
         Section section = Section.of(upStation, downStation, sectionRequest.getDistance());
 
-        Line line = lineComponent.getLineByIdOrThrow(lineId);
+        Line line = lineService.getLineByIdOrThrow(lineId);
         Sections sections = line.getSections();
 
         if (sections.getLastDownStationId() != upStation.getId()) {
@@ -55,16 +52,16 @@ public class SectionService {
 
         sections.setLastDownStationId(downStation.getId());
         sections.addSection(section);
-        lineComponent.saveLine(line);
+        lineService.saveLine(line);
 
         return new SectionResponse(lineId, section);
     }
 
     @Transactional
     public void deleteSection(Long lineId, long stationId) {
-        Section section = sectionComponent.getByDownStationId(stationId);
+        Section section = getByDownStationId(stationId);
 
-        Line line = lineComponent.getLineByIdOrThrow(lineId);
+        Line line = lineService.getLineByIdOrThrow(lineId);
         Sections sections = line.getSections();
 
         if (sections.getLastDownStationId() != stationId) {
@@ -78,8 +75,18 @@ public class SectionService {
         sections.removeSection(section);
         sections.setLastUpStationId(section.getUpStation().getId());
 
-        sectionComponent.deleteSection(section);
-        lineComponent.saveLine(line);
+        deleteSection(section);
+        lineService.saveLine(line);
+    }
+
+    public Section getByDownStationId(Long stationId) {
+        return sectionRepository.findByDownStationId(stationId).orElseThrow(
+                () -> new SectionException(String.valueOf(SECTION_NOT_FOUND))
+        );
+    }
+
+    public void deleteSection(Section section) {
+        sectionRepository.delete(section);
     }
 
 }
